requirements.txt
Django==1.8.3
django-tastypie==0.12.2
